Update SikuliSession.cs
Added DoubleClick methods. Tested and they work. I'm aware of bug report https://github.com/christianrondeau/SikuliSharp/issues/12 but I kept the timeout parameters in the methods for now, hope that's ok. Thanks for creating this .Net wrapper for Sikuli, it's the best of the three I tried to get working!